transient 생존주기 확인
다시 select 해올 때 초기화 되는 것으로 추정됨
boot component 삭제 및 likes 속도 향상을 위한 fetch정책 설정

+++ src/main/java/com/fashionlog/model/service/LikesService.java
@@ -7,9 +7,9 @@ import com.fashionlog.model.dto.Post;
 
 public interface LikesService {
 
-	public int doLike(Member member, Post post);
+	public Long doLike(Member member, Post post);
 
-	public int unLike(Member member, Post post) throws NoSuchElementException;
+	public Long unLike(Member member, Post post) throws NoSuchElementException;
 
 	public void countLikes();
 	public void countLikes(Post post);
